+++ test/Output.hs
@@ -32,9 +32,10 @@ tests = Tasty.testGroup "LLVM pretty-printing output tests"
         -- output are seen.
         s1, s2 :: Stmt
         s1 = Effect
-             (GEP True (Alias (Ident "hi")) (Typed Opaque dcu) [])
-             []
+             (GEP [GEP_Inbounds] (Alias (Ident "hi")) (Typed Opaque dcu) [])
+             mempty []
         s2 = Effect (Load PtrOpaque (Typed Opaque ValNull) Nothing Nothing)
+             mempty
              [ ("location", ValMdLoc $ DebugLoc { dlLine = 12
                                                 , dlCol = 34
                                                 , dlScope = ValMdRef 5
@@ -75,10 +76,10 @@ tests = Tasty.testGroup "LLVM pretty-printing output tests"
                                         }
         blk1 = BasicBlock { bbLabel = Just $ Named $ Ident "blk1"
                           , bbStmts =
-                            [ Result (Ident "r1") (Comment "insanity follows...") []
-                            , Effect (Jump $ Named $ Ident "blk1") []
-                            , Result (Ident "oh no") RetVoid []
-                            , Effect (Br (Typed (PrimType Metadata) ValZeroInit) (Anon 3) (Named "oh no")) []
+                            [ Result (Ident "r1") (Comment "insanity follows...") mempty []
+                            , Effect (Jump $ Named $ Ident "blk1") mempty []
+                            , Result (Ident "oh no") RetVoid mempty []
+                            , Effect (Br (Typed (PrimType Metadata) ValZeroInit) (Anon 3) (Named "oh no")) mempty []
                             ]
                           }
         blk2 = BasicBlock { bbLabel = Just $ Anon 123
@@ -149,6 +150,51 @@ tests = Tasty.testGroup "LLVM pretty-printing output tests"
   , testCase "Stmt 1, LLVM 11" $
     assertEqLines (ppToText $ ppLLVM 11 $ ppStmt s1) [sq|
       In LLVM 11, DICompileUnit adds rangesBaseAddress, sysroot, and sdk
+      ----
+      getelementptr inbounds %hi, opaque !DICompileUnit(language: 12,
+                                                        producer: "llvm-pretty-test",
+                                                        isOptimized: true,
+                                                        flags: "some flags",
+                                                        runtimeVersion: 3,
+                                                        emissionKind: 1,
+                                                        enums: !DITemplateTypeParameter(name: ttp),
+                                                        dwoId: 2,
+                                                        splitDebugInlining: false,
+                                                        debugInfoForProfiling: true,
+                                                        nameTableKind: 4,
+                                                        rangesBaseAddress: true,
+                                                        sysroot: "the root",
+                                                        sdk: "SDK")
+      ----
+      |]
+
+  , testCase "Stmt 1, LLVM 18" $
+    assertEqLines (ppToText $ ppLLVM 18 $ ppStmt s1) [sq|
+      Significant changes occur in LLVM 19; ensure that the output is as
+      expected just prior to those changes to properly capture the
+      inflection point.
+      ----
+      getelementptr inbounds %hi, opaque !DICompileUnit(language: 12,
+                                                        producer: "llvm-pretty-test",
+                                                        isOptimized: true,
+                                                        flags: "some flags",
+                                                        runtimeVersion: 3,
+                                                        emissionKind: 1,
+                                                        enums: !DITemplateTypeParameter(name: ttp),
+                                                        dwoId: 2,
+                                                        splitDebugInlining: false,
+                                                        debugInfoForProfiling: true,
+                                                        nameTableKind: 4,
+                                                        rangesBaseAddress: true,
+                                                        sysroot: "the root",
+                                                        sdk: "SDK")
+      ----
+      |]
+
+  , testCase "Stmt 1, LLVM 19" $
+    assertEqLines (ppToText $ ppLLVM 19 $ ppStmt s1) [sq|
+      In LLVM 19, the GEP instruction "inbounds" is no longer a boolean
+      but a flag with multiple possible values (only one is checked here).
       ----
       getelementptr inbounds %hi, opaque !DICompileUnit(language: 12,
                                                         producer: "llvm-pretty-test",